requirements.txt
streamlit>=1.32.0
tensorflow>=2.12.0
Pillow>=9.0.0
numpy>=1.23.0
